+++ eq-author-api/schema/resolvers/utils/index.js
@@ -1,16 +1,18 @@
-const getters = require("./getters");
+const logicGetters = require("./logicGetters");
 const sectionGetters = require("./sectionGetters");
 const folderGetters = require("./folderGetters");
 const pageGetters = require("./pageGetters");
+const answerGetters = require("./answerGetters");
 const helpers = require("./helpers");
 const theme = require("./theme");
 const validation = require("./validation");
 
 module.exports = {
-  ...getters,
+  ...logicGetters,
   ...sectionGetters,
-  ...pageGetters,
   ...folderGetters,
+  ...pageGetters,
+  ...answerGetters,
   ...helpers,
   ...theme,
   ...validation,
